refactor: class ResultService
LottoService의 역할 분리
